refactor: 계층 변경
